Move `ObjectUsage` to be a subclass
